💥 Parsers: Remove legacy code
Remove several "legacy" modules with outdated code based on previous Quantum ESPRESSO
< v6.2 XML schemas that we no longer support.

The following functions are moved from `parse_xml/pw/legacy.py` into the `parse_xml/parse.py` module:

- `read_xml_card`
- `parse_xml_child_integer`
- `parse_xml_child_bool`
- `str2bool`

Since they are still used in non-legacy parts of the code base. Judging their logic and
whether they can be replaced is beyond the scope of this commit.

From the `versions.py` module, we only really still need the functionality to obtain
the schema filepath. This logic is gathered into a single `get_schema_filepath` function
which is moved to the `parse_xml/parse.py` module.

- For the `PwParser`, the `parser_options` were passed to the `.parse_xml` method, but
  not used. Hence we remove them from the signature.
- The `default` `cp.x` test still used the old XML format in `data-file.xml`, and the
  remaining tests provide sufficient coverage. Hence we can safely remove it.
- the `dir_with_bands` input is no longer necessary, hence we remove any references to
  it. This also means we no longer need to try to obtain the retrieved temporary folder
  for the `PwParser`.
