+++ src/aiida_quantumespresso/parsers/parse_xml/parse.py
@@ -1,6 +1,8 @@
 import collections
 import contextlib
-from urllib.error import URLError
+from pathlib import Path
+from xml.etree import ElementTree
+from xml.dom.minidom import Element
 
 import numpy as np
 from packaging.version import Version
@@ -8,9 +10,9 @@ from qe_tools import CONSTANTS
 from xmlschema import XMLSchema
 
 from aiida_quantumespresso.utils.mapping import get_logging_container
+from aiida_quantumespresso.parsers import QEOutputParsingError
 
-from .exceptions import XMLParseError
-from .versions import get_default_schema_filepath, get_schema_filepath
+from .exceptions import XMLParseError, XMLUnsupportedFormatError
 
 
 def raise_parsing_error(message):
@@ -40,7 +42,93 @@ def cell_volume(a1, a2, a3):
     return abs(float(a1[0] * a_mid_0 + a1[1] * a_mid_1 + a1[2] * a_mid_2))
 
 
-def parse_xml_post_6_2(xml):
+def read_xml_card(dom, cardname):
+    try:
+        root_node = [_ for _ in dom.childNodes if isinstance(_, Element) and _.nodeName == 'Root'][0]
+        the_card = [_ for _ in root_node.childNodes if _.nodeName == cardname][0]
+        # the_card = dom.getElementsByTagName(cardname)[0]
+        return the_card
+    except Exception as e:
+        print(e)
+        raise QEOutputParsingError(f'Error parsing tag {cardname}')
+
+
+def parse_xml_child_integer(tagname, target_tags):
+    try:
+        # a=target_tags.getElementsByTagName(tagname)[0]
+        a = [_ for _ in target_tags.childNodes if _.nodeName == tagname][0]
+        b = a.childNodes[0]
+        return int(b.data)
+    except Exception:
+        raise QEOutputParsingError(f'Error parsing tag {tagname} inside {target_tags.tagName}')
+
+
+def parse_xml_child_bool(tagname, target_tags):
+    try:
+        # a=target_tags.getElementsByTagName(tagname)[0]
+        a = [_ for _ in target_tags.childNodes if _.nodeName == tagname][0]
+        b = a.childNodes[0]
+        return str2bool(b.data)
+    except Exception:
+        raise QEOutputParsingError(f'Error parsing tag {tagname} inside {target_tags.tagName}')
+
+
+def str2bool(string):
+    try:
+        false_items = ['f', '0', 'false', 'no']
+        true_items = ['t', '1', 'true', 'yes']
+        string = str(string.lower().strip())
+        if string in false_items:
+            return False
+        if string in true_items:
+            return True
+        else:
+            raise QEOutputParsingError(f'Error converting string {string} to boolean value.')
+    except Exception:
+        raise QEOutputParsingError('Error converting string to boolean.')
+
+
+def get_schema_filepath(xml):
+    """Return the absolute filepath to the XML schema file that can be used to parse the given XML.
+
+    :param xml: the pre-parsed XML object
+    :return: the XSD absolute filepath
+    """
+    xml_schema_location_key = '{http://www.w3.org/2001/XMLSchema-instance}schemaLocation'
+    # The part in curly brackets is an expanded namespace
+    dir_path_schemas = Path(__file__).resolve().parent / 'schemas'
+
+    element_root = xml.getroot()
+
+    if element_root is None:
+        raise XMLParseError('Could not find the XML root!')
+
+    # Patch for QE v7.0: The scheme file name was not updated in the `xsi.schemaLocation` element
+    with contextlib.suppress(AttributeError):
+        if element_root.find('general_info').find('creator').get('VERSION') == '7.0':
+            return dir_path_schemas / 'qes_211101.xsd'
+
+    element_schema_location = element_root.get(xml_schema_location_key)
+    # e.g. "http://www.quantum-espresso.org/ns/qes/qes-1.0 http://www.quantum-espresso.org/ns/qes/qes-1.0.xsd"
+
+    if element_schema_location is None:
+        raise XMLParseError('Could not find the schema location!')
+
+    schema_location = element_schema_location.split()[1]  # e.g. "http://www.quantum-espresso.org/ns/qes/qes-1.0.xsd"
+    schema_filename = schema_location.rpartition('/')[2]  # e.g. "qes-1.0.xsd"
+
+    schema_filepath = dir_path_schemas / schema_filename
+
+    if not schema_filepath.exists():
+        raise XMLUnsupportedFormatError(
+            f'Cannot find schema {schema_filepath.name} in {dir_path_schemas}.\n'
+            'Make sure you are running a supported Quantum ESPRESSO version.'
+        )
+
+    return schema_filepath
+
+
+def parse_xml(xml_file):
     """Parse the content of XML output file written by `pw.x` and `cp.x` with the new schema-based XML format.
 
     :param xml: parsed XML
@@ -50,22 +138,13 @@ def parse_xml_post_6_2(xml):
 
     logs = get_logging_container()
 
-    schema_filepath = get_schema_filepath(xml)
-
     try:
-        xsd = XMLSchema(schema_filepath)
-    except URLError:
-        # If loading the XSD file specified in the XML file fails, we try the default
-        schema_filepath_default = get_default_schema_filepath()
+        xml_parsed = ElementTree.parse(xml_file)
+    except ElementTree.ParseError as exception:
+        raise XMLParseError('error while parsing XML file') from exception
 
-        try:
-            xsd = XMLSchema(schema_filepath_default)
-        except URLError:
-            raise XMLParseError(
-                f'Could not open or parse the XSD files {schema_filepath} and {schema_filepath_default}'
-            )
-        else:
-            schema_filepath = schema_filepath_default
+    schema_filepath = get_schema_filepath(xml_parsed)
+    xsd = XMLSchema(schema_filepath)
 
     # Validate XML document against the schema
     # Returned dictionary has a structure where, if tag ['key'] is "simple", xml_dictionary['key'] returns its content.
@@ -75,7 +154,7 @@ def parse_xml_post_6_2(xml):
     #  xml_dictionary['key']['@attr'] returns its attribute 'attr'
     #  xml_dictionary['key']['nested_key'] goes one level deeper.
 
-    xml_dictionary, errors = xsd.to_dict(xml, validation='lax')
+    xml_dictionary, errors = xsd.to_dict(xml_parsed, validation='lax')
     if errors:
         logs.error.append(f'{len(errors)} XML schema validation error(s) schema: {schema_filepath}:')
         for err in errors:
