--- tests/parsers/fixtures/cp/default/aiida.cel
@@ -1,40 +1,0 @@
-      1  0.00007257
-    0.00000000    5.10226056    5.10226056
-    5.10226056    0.00000000    5.10226056
-    5.10226056    5.10226056    0.00000000
-      2  0.00014513
-    0.00000000    5.10226056    5.10226056
-    5.10226056    0.00000000    5.10226056
-    5.10226056    5.10226056    0.00000000
-      3  0.00021770
-    0.00000000    5.10226056    5.10226056
-    5.10226056    0.00000000    5.10226056
-    5.10226056    5.10226056    0.00000000
-      4  0.00029027
-    0.00000000    5.10226056    5.10226056
-    5.10226056    0.00000000    5.10226056
-    5.10226056    5.10226056    0.00000000
-      5  0.00036283
-    0.00000000    5.10226056    5.10226056
-    5.10226056    0.00000000    5.10226056
-    5.10226056    5.10226056    0.00000000
-      6  0.00043540
-    0.00000000    5.10226056    5.10226056
-    5.10226056    0.00000000    5.10226056
-    5.10226056    5.10226056    0.00000000
-      7  0.00050797
-    0.00000000    5.10226056    5.10226056
-    5.10226056    0.00000000    5.10226056
-    5.10226056    5.10226056    0.00000000
-      8  0.00058053
-    0.00000000    5.10226056    5.10226056
-    5.10226056    0.00000000    5.10226056
-    5.10226056    5.10226056    0.00000000
-      9  0.00065310
-    0.00000000    5.10226056    5.10226056
-    5.10226056    0.00000000    5.10226056
-    5.10226056    5.10226056    0.00000000
-     10  0.00072567
-    0.00000000    5.10226056    5.10226056
-    5.10226056    0.00000000    5.10226056
-    5.10226056    5.10226056    0.00000000
