--- tests/parsers/fixtures/cp/default/aiida.eig
@@ -1,30 +1,0 @@
-  STEP:      1       0.00
-   Eigenvalues (eV), kp =   1 , spin =  1
-   31.79   35.51   46.15   53.94
-  STEP:      2       0.00
-   Eigenvalues (eV), kp =   1 , spin =  1
-   27.59   31.10   41.63   50.68
-  STEP:      3       0.00
-   Eigenvalues (eV), kp =   1 , spin =  1
-   22.33   25.45   35.87   46.64
-  STEP:      4       0.00
-   Eigenvalues (eV), kp =   1 , spin =  1
-   17.14   19.69   29.77   42.45
-  STEP:      5       0.00
-   Eigenvalues (eV), kp =   1 , spin =  1
-   12.78   14.63   23.88   38.46
-  STEP:      6       0.00
-   Eigenvalues (eV), kp =   1 , spin =  1
-    9.55   10.74   18.61   34.84
-  STEP:      7       0.00
-   Eigenvalues (eV), kp =   1 , spin =  1
-    7.45    8.16   14.24   31.63
-  STEP:      8       0.00
-   Eigenvalues (eV), kp =   1 , spin =  1
-    6.25    6.74   10.97   28.79
-  STEP:      9       0.00
-   Eigenvalues (eV), kp =   1 , spin =  1
-    5.74    5.99    9.04   26.28
-  STEP:     10       0.00
-   Eigenvalues (eV), kp =   1 , spin =  1
-    5.15    6.01    8.46   24.02
