--- tests/parsers/fixtures/cp/default/aiida.evp
@@ -1,11 +1,0 @@
-#   nfi    time(ps)        ekinc        T_cell(K)     Tion(K)          etot               enthal               econs               econt          Volume        Pressure(GPa)        EXX               EVDW
-      1  7.256653E-05  3.436858E-01  0.000000E+00  0.000000E+00          3.98558847          3.98558847          3.98558847          4.32927424  2.656549E+02      -0.00000
-      2  1.451331E-04  9.667654E-01  0.000000E+00  0.000000E+00          3.08755201          3.08755201          3.08755201          4.05431736  2.656549E+02      -0.00000
-      3  2.176996E-04  1.707442E+00  0.000000E+00  0.000000E+00          1.79358149          1.79358149          1.79358149          3.50102300  2.656549E+02      -0.00000
-      4  2.902661E-04  2.391118E+00  0.000000E+00  0.000000E+00          0.26880457          0.26880457          0.26880457          2.65992241  2.656549E+02      -0.00000
-      5  3.628326E-04  2.901277E+00  0.000000E+00  0.000000E+00         -1.32073449         -1.32073449         -1.32073449          1.58054234  2.656549E+02      -0.00000
-      6  4.353992E-04  3.176072E+00  0.000000E+00  0.000000E+00         -2.82798368         -2.82798368         -2.82798368          0.34808793  2.656549E+02      -0.00000
-      7  5.079657E-04  3.206193E+00  0.000000E+00  0.000000E+00         -4.14531151         -4.14531151         -4.14531151         -0.93911888  2.656549E+02      -0.00000
-      8  5.805322E-04  3.026046E+00  0.000000E+00  0.000000E+00         -5.21561573         -5.21561573         -5.21561573         -2.18957000  2.656549E+02      -0.00000
-      9  6.530988E-04  2.697539E+00  0.000000E+00  0.000000E+00         -6.03008800         -6.03008800         -6.03008800         -3.33254933  2.656549E+02      -0.00000
-     10  7.256653E-04  2.291610E+00  0.000000E+00  0.000000E+00         -6.61628032         -6.61628032         -6.61628032         -4.32467025  2.656549E+02      -0.00000
