--- tests/parsers/fixtures/cp/default/aiida.nos
@@ -1,10 +1,0 @@
-      1  7.256653E-05      0.00000000      0.00000000      0.00000000      0.00000000      0.00000000      0.00000000
-      2  1.451331E-04      0.00000000      0.00000000      0.00000000      0.00000000      0.00000000      0.00000000
-      3  2.176996E-04      0.00000000      0.00000000      0.00000000      0.00000000      0.00000000      0.00000000
-      4  2.902661E-04      0.00000000      0.00000000      0.00000000      0.00000000      0.00000000      0.00000000
-      5  3.628326E-04      0.00000000      0.00000000      0.00000000      0.00000000      0.00000000      0.00000000
-      6  4.353992E-04      0.00000000      0.00000000      0.00000000      0.00000000      0.00000000      0.00000000
-      7  5.079657E-04      0.00000000      0.00000000      0.00000000      0.00000000      0.00000000      0.00000000
-      8  5.805322E-04      0.00000000      0.00000000      0.00000000      0.00000000      0.00000000      0.00000000
-      9  6.530988E-04      0.00000000      0.00000000      0.00000000      0.00000000      0.00000000      0.00000000
-     10  7.256653E-04      0.00000000      0.00000000      0.00000000      0.00000000      0.00000000      0.00000000
