--- tests/parsers/fixtures/cp/default/aiida.pos
@@ -1,30 +1,0 @@
-      1  0.00007257
-     0.00000000000000E+00     0.00000000000000E+00     0.00000000000000E+00
-     0.25511302793956E+01     0.25511302793956E+01     0.25511302793956E+01
-      2  0.00014513
-     0.00000000000000E+00     0.00000000000000E+00     0.00000000000000E+00
-     0.25511302793956E+01     0.25511302793956E+01     0.25511302793956E+01
-      3  0.00021770
-     0.00000000000000E+00     0.00000000000000E+00     0.00000000000000E+00
-     0.25511302793956E+01     0.25511302793956E+01     0.25511302793956E+01
-      4  0.00029027
-     0.00000000000000E+00     0.00000000000000E+00     0.00000000000000E+00
-     0.25511302793956E+01     0.25511302793956E+01     0.25511302793956E+01
-      5  0.00036283
-     0.00000000000000E+00     0.00000000000000E+00     0.00000000000000E+00
-     0.25511302793956E+01     0.25511302793956E+01     0.25511302793956E+01
-      6  0.00043540
-     0.00000000000000E+00     0.00000000000000E+00     0.00000000000000E+00
-     0.25511302793956E+01     0.25511302793956E+01     0.25511302793956E+01
-      7  0.00050797
-     0.00000000000000E+00     0.00000000000000E+00     0.00000000000000E+00
-     0.25511302793956E+01     0.25511302793956E+01     0.25511302793956E+01
-      8  0.00058053
-     0.00000000000000E+00     0.00000000000000E+00     0.00000000000000E+00
-     0.25511302793956E+01     0.25511302793956E+01     0.25511302793956E+01
-      9  0.00065310
-     0.00000000000000E+00     0.00000000000000E+00     0.00000000000000E+00
-     0.25511302793956E+01     0.25511302793956E+01     0.25511302793956E+01
-     10  0.00072567
-     0.00000000000000E+00     0.00000000000000E+00     0.00000000000000E+00
-     0.25511302793956E+01     0.25511302793956E+01     0.25511302793956E+01
