--- tests/parsers/fixtures/cp/default/aiida.vel
@@ -1,30 +1,0 @@
-      1  0.00007257
-     0.00000000000000E+00     0.00000000000000E+00     0.00000000000000E+00
-     0.00000000000000E+00     0.00000000000000E+00     0.00000000000000E+00
-      2  0.00014513
-     0.00000000000000E+00     0.00000000000000E+00     0.00000000000000E+00
-     0.00000000000000E+00     0.00000000000000E+00     0.00000000000000E+00
-      3  0.00021770
-     0.00000000000000E+00     0.00000000000000E+00     0.00000000000000E+00
-     0.00000000000000E+00     0.00000000000000E+00     0.00000000000000E+00
-      4  0.00029027
-     0.00000000000000E+00     0.00000000000000E+00     0.00000000000000E+00
-     0.00000000000000E+00     0.00000000000000E+00     0.00000000000000E+00
-      5  0.00036283
-     0.00000000000000E+00     0.00000000000000E+00     0.00000000000000E+00
-     0.00000000000000E+00     0.00000000000000E+00     0.00000000000000E+00
-      6  0.00043540
-     0.00000000000000E+00     0.00000000000000E+00     0.00000000000000E+00
-     0.00000000000000E+00     0.00000000000000E+00     0.00000000000000E+00
-      7  0.00050797
-     0.00000000000000E+00     0.00000000000000E+00     0.00000000000000E+00
-     0.00000000000000E+00     0.00000000000000E+00     0.00000000000000E+00
-      8  0.00058053
-     0.00000000000000E+00     0.00000000000000E+00     0.00000000000000E+00
-     0.00000000000000E+00     0.00000000000000E+00     0.00000000000000E+00
-      9  0.00065310
-     0.00000000000000E+00     0.00000000000000E+00     0.00000000000000E+00
-     0.00000000000000E+00     0.00000000000000E+00     0.00000000000000E+00
-     10  0.00072567
-     0.00000000000000E+00     0.00000000000000E+00     0.00000000000000E+00
-     0.00000000000000E+00     0.00000000000000E+00     0.00000000000000E+00
