--- tests/parsers/fixtures/cp/default/print_counter.xml
@@ -1,12 +1,0 @@
-<?xml version="1.0"?>
-<?iotk version="1.2.0"?>
-<?iotk file_version="1.0"?>
-<?iotk binary="F"?>
-<?iotk qe_syntax="F"?>
-<PRINT_COUNTER>
-  <LAST_SUCCESSFUL_PRINTOUT>
-    <STEP type="integer" size="1">
-        10
-    </STEP>
-  </LAST_SUCCESSFUL_PRINTOUT>
-</PRINT_COUNTER>
